+++ tutor/utils.py
@@ -130,7 +130,7 @@ def get_user_id():
     """
     Portable way to get user ID. Note: I have no idea if it actually works on windows...
     """
-    if sys.platform == "windows":
+    if sys.platform == "win32":
         return int(check_output("id", "-u").decode())
     return os.getuid()