+++ tests/test_env.py
@@ -28,9 +28,9 @@ class EnvTests(unittest.TestCase):
     def test_is_binary_file(self):
         self.assertTrue(env.is_binary_file("/home/somefile.ico"))
 
-    def test_find_path(self):
+    def test_find_os_path(self):
         renderer = env.Renderer({}, [env.TEMPLATES_ROOT])
-        path = renderer.find_path("local/docker-compose.yml")
+        path = renderer.find_os_path("local/docker-compose.yml")
         self.assertTrue(os.path.exists(path))
 
     def test_pathjoin(self):