+++ tutor/utils.py
@@ -1,11 +1,4 @@
 import base64
-from sys import platform
-if platform == "win32":
-    from pcrypt import crypt
-else:
-    from crypt import crypt
-
-from hmac import compare_digest
 import json
 import os
 import random
@@ -17,6 +10,7 @@ import sys
 
 import click
 from Crypto.PublicKey import RSA
+from Crypto.Protocol.KDF import bcrypt, bcrypt_check
 
 from . import exceptions
 from . import fmt
@@ -24,20 +18,23 @@ from . import fmt
 
 def encrypt(text):
     """
-    Encrypt some textual content. The method employed is the same as suggested in the
-    `python docs <https://docs.python.org/3/library/crypt.html#examples>`__. The
-    encryption process is compatible with the password verification performed by
+    Encrypt some textual content with bcrypt.
+    https://pycryptodome.readthedocs.io/en/latest/src/protocol/kdf.html#bcrypt
+    The encryption process is compatible with the password verification performed by
     `htpasswd <https://httpd.apache.org/docs/2.4/programs/htpasswd.html>`__.
     """
-    hashed = crypt(text)
-    return crypt(text, hashed)
+    return bcrypt(text.encode(), 12).decode()
 
 
 def verify_encrypted(encrypted, text):
     """
     Return True/False if the encrypted content corresponds to the unencrypted text.
     """
-    return compare_digest(crypt(text, encrypted), encrypted)
+    try:
+        bcrypt_check(text.encode(), encrypted.encode())
+        return True
+    except ValueError:
+        return False
 
 
 def ensure_file_directory_exists(path):
@@ -129,6 +126,15 @@ def walk_files(path):
             yield os.path.join(dirpath, filename)
 
 
+def get_user_id():
+    """
+    Portable way to get user ID. Note: I have no idea if it actually works on windows...
+    """
+    if sys.platform == "windows":
+        return int(check_output("id", "-u").decode())
+    return os.getuid()
+
+
 def docker_run(*command):
     args = ["run", "--rm"]
     if is_a_tty():
@@ -193,6 +199,5 @@ def check_output(*command):
     click.echo(fmt.command(" ".join(command)))
     try:
         return subprocess.check_output(command)
-    except:
+    finally:
         fmt.echo_error("Command failed: {}".format(" ".join(command)))
-        raise